requirements.txt
streamlit
google-genai
pymupdf
numpy
python-dotenv